feat : 여행 리포트 생성 API

+++ trippy-server/src/main/java/org/scoula/controller/travel/log/dto/res/TravelLogDTO.java
@@ -11,5 +11,5 @@ public record TravelLogDTO(
         String destination,
         Boolean isGenerated,
         String travelImg,
-        Integer memberCount
+        Long memberCount
 ) {}
